Refcator the genirics/variadics for size
Added the variadics to the index members to allow the resolver
to correctly address them / modify them once a generic type needs to be
replaced
