+++ src/typesystem.rs
@@ -116,6 +116,14 @@ impl VarArgs {
     pub fn is_sized(&self) -> bool {
         matches!(self, VarArgs::Sized(..))
     }
+
+    pub fn as_typed(&self, new_type: &str) -> VarArgs {
+        match self {
+            VarArgs::Sized(Some(_)) => VarArgs::Sized(Some(new_type.to_string())),
+            VarArgs::Unsized(Some(_)) => VarArgs::Unsized(Some(new_type.to_string())),
+            _ => self.clone(),
+        }
+    }
 }
 
 #[derive(Debug, Clone, PartialEq)]
@@ -185,7 +193,6 @@ pub enum DataTypeInformation {
     Struct {
         name: TypeId,
         member_names: Vec<String>,
-        varargs: Option<VarArgs>,
         source: StructSource,
     },
     Array {
@@ -320,40 +327,6 @@ impl DataTypeInformation {
                 | DataTypeInformation::Float { .. }
                 | &DataTypeInformation::Enum { .. } // internally an enum is represented as a DINT
         )
-    }
-
-    pub fn is_variadic(&self) -> bool {
-        matches!(
-            self,
-            DataTypeInformation::Struct {
-                varargs: Some(_),
-                ..
-            }
-        )
-    }
-
-    pub fn get_variadic(&self) -> Option<&VarArgs> {
-        if let DataTypeInformation::Struct { varargs, .. } = &self {
-            varargs.as_ref()
-        } else {
-            None
-        }
-    }
-
-    pub fn get_variadic_type(&self) -> Option<&str> {
-        if let DataTypeInformation::Struct {
-            varargs: Some(inner_type),
-            ..
-        } = &self
-        {
-            let inner_type = match inner_type {
-                VarArgs::Sized(t) => t,
-                VarArgs::Unsized(t) => t,
-            };
-            inner_type.as_ref().map(String::as_str)
-        } else {
-            None
-        }
     }
 
     pub fn is_generic(&self, index: &Index) -> bool {
